# Conflicts:
#	src/main/java/com/git/branch/mastertest.java

+++ src/main/java/com/git/branch/mastertest.java
@@ -3,5 +3,5 @@ package com.git.branch;
 public class mastertest {
     Integer num = 666;
     int num2 = 888;
-    Long num3 = 7777L;
+    Double num3 = 7777.00;
 }
